gpytorch.settings.variational_cholesky_jitter can be set dynamically.
Previously, this context manager was only used when VariationalStrategy
modules were initialized. With this change,
gpytorch.settings.variational_cholesky_jitter will dynamically change
the jitter value (for variational models already in use), unless the
user specifies a `jitter_val` in the VariationalStrategy constructor.

In addition, this PR adds type hintsd to a majority of the
VariationalStrategy modules.

[Fixes #2244]